requirements.txt
sentence-transformers>=2.5.1
numpy>=1.26.4
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
requests>=2.31.0
streamlit>=1.25.0
opencv-python>=4.9.0.74
pytesseract>=0.3.10
Pillow>=10.0.0
piexif>=1.1.3
stegano>=0.1.4